Moving refined_string_type to util
This is a more appropriate location for this module since it's used
both in the preprocessing of goto-programs and the string solver.
